requirements.txt
argcomplete==3.5.3
babel==2.17.0
Beaker==1.12.1
beautifulsoup4==4.12.3
blivet==3.11.0
blivet-gui==2.6.0
Brlapi==0.8.5
certifi==2025.1.31
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
contourpy==1.3.1
crypt_r==3.13.1
cryptography==43.0.0
cupshelpers==1.0
cycler==0.12.1
dasbus==1.7
dbus-python==1.3.2
distro==1.9.0
dnf==4.22.0
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.17.0
fonttools==4.56.0
fros==1.1
fsspec==2025.2.0
humanize==4.9.0
idna==3.7
imageio==2.37.0
iso639==0.1.4
Jinja2==3.1.6
kiwisolver==1.4.8
langtable==0.0.68
lazy_loader==0.4
libcomps==0.1.21
libdnf==0.73.4
louis==3.28.0
lxml==5.2.1
Mako==1.2.3
MarkupSafe==2.1.5
matplotlib==3.10.1
mpmath==1.3.0
navms_particle_filter @ git+ssh://git@github.com/NAVMS-FURG/particleFilter.git@09a66328db22785209004a8a75d217c6a4a2bd6f
networkx==3.4.2
nftables==0.1
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
olefile==0.47
opencv-python==4.11.0.86
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
Paste==3.10.1
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.0.0
ply==3.11
productmd==1.43
progressbar2==3.53.2
ptyprocess==0.7.0
pwquality==1.4.5
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
pyenchant==3.2.2
pygit2==1.17.0
PyGObject==3.48.2
pyinotify==0.9.6
pykickstart==3.58
pyOpenSSL==24.2.1
pyparsing==3.2.1
pyparted==3.13.0
pyproj==3.7.1
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
python-utils==3.7.0
pytz==2025.1
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.1
regex==2024.9.11
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
rpm==4.20.0
rpmautospec==0.7.3
rpmautospec-core==0.1.5
scikit-image==0.25.2
scipy==1.15.2
selinux @ file:///builddir/build/BUILD/libselinux-3.7-build/libselinux-3.7/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.7-build/selinux-3.7/python/sepolicy
setools==4.5.1
setuptools==69.2.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sos==4.8.2
soupsieve==2.6
sympy==1.13.1
systemd-python==235
Tempita==0.5.2
tifffile==2025.2.18
torch==2.5.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
xkbregistry==0.3
